Refactor the disk processor into smaller chunks
